+++ protostar/test_runner/test_context.py
@@ -1,8 +1,8 @@
 from types import SimpleNamespace
 from typing import Any
 
-from protostar.starknet.hint_local import HintLocal
 from protostar.commands.test.test_environment_exceptions import SimpleReportedException
+from protostar.starknet.hint_local import HintLocal
 
 
 class TestContext(SimpleNamespace):
