+++ protostar/test_runner/test_context_test.py
@@ -1,7 +1,8 @@
 import pytest
 
-from protostar.commands.test.test_context import TestContext
 from protostar.commands.test.test_environment_exceptions import ReportedException
+
+from .test_context import TestContext
 
 
 def test_immutability():
